Start a component lifecycle guide
